refactor: move settings dialog to global dialogs
