+++ frontend/src/components/editor-page/app-bar/app-bar.tsx
@@ -8,9 +8,9 @@ import { useOutlineButtonVariant } from '../../../hooks/dark-mode/use-outline-bu
 import { CheatsheetButton } from '../../cheatsheet/cheatsheet-button'
 import { NewNoteButton } from '../../common/new-note-button/new-note-button'
 import { ShowIf } from '../../common/show-if/show-if'
+import { SettingsButton } from '../../global-dialogs/settings-dialog/settings-button'
 import { SignInButton } from '../../landing-layout/navigation/sign-in-button'
 import { UserDropdown } from '../../landing-layout/navigation/user-dropdown'
-import { SettingsButton } from '../../layout/settings-dialog/settings-button'
 import { HelpButton } from './help-button/help-button'
 import { NavbarBranding } from './navbar-branding'
 import { ReadOnlyModeButton } from './read-only-mode-button'
