+++ frontend/src/components/global-dialogs/settings-dialog/global/available-languages.ts
