+++ frontend/src/components/global-dialogs/settings-dialog/utils/on-off-button-group.spec.tsx
